// src/horreum/generated/models/KeycloakConfig.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type KeycloakConfig = {
  /**
   * Keycloak realm securing Horreum instance
   */
  realm?: string;
  /**
   * URL of Keycloak instance securing Horreum
   */
  url?: string;
  /**
   * Keycloak client ID in Horreum realm for User Interface
   */
  clientId?: string;
};